+++ models/Hacker.cs
@@ -6,7 +6,7 @@ namespace heistTwo
     {
         public string Name { get; set; }
         public int SkillLevel { get; set; }
-        public int PercentageCut { get; set; }
+        public double PercentageCut { get; set; }
         public void PerformSkill(Bank bank)
         {
             bank.AlarmScore -= SkillLevel;
